:sparkles: Refactor RetrieveConfigWizard to be inline with DiscoverImportWizard
